bloginfo Ej 4.21 - paso 9 listo

+++ controllers/blogs.js
@@ -50,8 +50,43 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id);
-    response.status(204).end();
+    try {
+        // verificar token
+        const decodedToken = jwt.verify(request.token, process.env.SECRET);
+        if (!decodedToken.id) {
+            return response.status(401).json({ error: "token invalid" });
+        }
+
+        // verificar usuario
+        const user = await User.findById(decodedToken.id);
+        if (!user) {
+            return response.status(404).json({ error: "user not found" });
+        }
+
+        // verificar si el blog existe
+        const blog = await Blog.findById(request.params.id);
+        if (!blog) {
+            return response.status(404).json({ error: "blog not found" });
+        }
+
+        // verificar si el usuario es el propietario del blog
+        if (blog.user.toString() === user._id.toString()) {
+            await Blog.findByIdAndDelete(request.params.id);
+            return response.status(204).end();
+        } else {
+            return response
+                .status(401)
+                .json({
+                    error: "unauthorized: only the creator can delete this blog",
+                });
+        }
+    } catch (error) {
+        if (error.name === "JsonWebTokenError") {
+            return response.status(401).json({ error: "token invalid" });
+        }
+
+        return response.status(500).json({ error: "internal server error" });
+    }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
